Change the checkpointing logic to reuse reclaim
Checkpointing depends on the file being switched before
the checkpointing is started but that was never happening
because reclaim would only occur when the file was full
(and not when the file was forced to roll over).

This fixes that by embedding the checkpointing logic in
the pre-existing roll-over logic.