Allow to obtain the number of pending checkpoints
This is a measure of backpressure: if the number of checkpoints
pendings increases too much it means we're falling behind.

This also make it so that we panic!() when a checkpoint fails.
It's important to fail here because most of the time we don't
know how to recover from a checkpoint having failed.