No more panics on errors & allow to inspect whether the checkpointing thread is running (#8)
This adds a couple of things, one is to stop the checkpoiting thread from
panicking and instead allow it to just return the error. This will help
in debugging cases where there are many checkpointing threads.

The second one is to add a method that allows to inspect whether the
checkpointing thread is running.